requirements.txt
openai
streamlit
loguru
PyYAML
mysql-connector-python